third_party/WebKit/LayoutTests/html/dialog/modal-dialog-ancestor-is-inert-expected.txt

Test that ancestors of modal <dialog> are inert. To test manually, click the left box. There should be no change. Then click the right box. If both boxes turn green, the test passes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Clicking on ancestor
PASS handledEvent.document is true
PASS handledEvent.body is false
PASS handledEvent.dialog is false
PASS handledEvent.div is false
Clicking on dialog
PASS handledEvent.document is true
PASS handledEvent.body is true
PASS handledEvent.dialog is true
PASS handledEvent.div is true
PASS successfullyParsed is true

TEST COMPLETE
